+++ src/command/sample_encode.rs
@@ -294,6 +294,7 @@ pub fn run(
                                 .max(input_pixel_format.unwrap_or(PixelFormat::Yuv444p10le)),
                             args.vfilter.as_deref(),
                         ),
+                        vmaf.vmaf_fps,
                     )?;
                     let mut vmaf = pin!(vmaf);
                     let mut logger = ProgressLogger::new("ab_av1::vmaf", Instant::now());